task: Code review changes
Create a MissionTestHelper for testing mission-related functionality

ref #2072
