+++ mars-sim-core/src/test/java/com/mars_sim/core/mission/steps/MissionBoardVehicleStepTest.java
@@ -6,54 +6,20 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import java.util.ArrayList;
-
 import org.junit.jupiter.api.Test;
 
-import com.mars_sim.core.MarsSimContext;
-import com.mars_sim.core.map.location.LocalPosition;
-import com.mars_sim.core.mission.MetaMission;
+import com.mars_sim.core.mission.MissionTestHelper;
 import com.mars_sim.core.mission.MissionVehicleProject;
 import com.mars_sim.core.person.ai.mission.MissionType;
-import com.mars_sim.core.person.ai.task.util.Worker;
-import com.mars_sim.core.structure.Settlement;
 import com.mars_sim.core.test.MarsSimUnitTest;
 
-public class MissionBoardVehicleStepTest extends MarsSimUnitTest {
-
-    /**
-     * Build a Roster for a mission with the specified number of workers and a vehicle.
-     * 
-     * @param context The simulation context
-     * @param s The settlement where the mission is based
-     * @param workerNum The number of workers to include in the roster
-     * @param roverType The type of rover to use for the mission
-     * @param onVehicle Whether the workers should be placed on the vehicle initially
-     * @return A MetaMission.Roster containing the leader, members, and vehicle
-     */
-    public static MetaMission.Roster buildRoster(MarsSimContext context, Settlement s, int workerNum,
-                            String roverType, boolean onVehicle) {
-        var r = context.buildRover(s, "Rover", LocalPosition.DEFAULT_POSITION, roverType);
-        
-        var l = context.buildPerson("Leader", s);
-        var workers  = new ArrayList<Worker>();
-        for(var i = 0; i < workerNum; i++) {
-             workers.add(context.buildPerson("Worker " + i, s));
-        }
-
-        if (onVehicle) {
-            workers.forEach(w -> w.transfer(r));
-            l.transfer(r);
-        }
-
-        return new MetaMission.Roster(l, workers, r);
-    }
+class MissionBoardVehicleStepTest extends MarsSimUnitTest {
 
     @Test
     void testNoEjectLeader() {
         var s = buildSettlement("Test 1");
 
-        var r = buildRoster(getContext(), s, 2, EXPLORER_ROVER, false);
+        var r = MissionTestHelper.buildRoster(getContext(), s, 2, EXPLORER_ROVER, false);
 
 		var project = new MissionVehicleProject(null, MissionType.VISIT_LANDMARK, 10, r);
         var st = new MissionBoardVehicleStep(project);
@@ -73,7 +39,7 @@ public class MissionBoardVehicleStepTest extends MarsSimUnitTest {
         
         project.execute(l);
 
-        // Push time to simulate the worker missing mission
+        // Push time to simulate the leader missing mission
         var clock = getSim().getMasterClock();
         var startTime = clock.getMarsTime().addTime(MissionBoardVehicleStep.DEPARTURE_DURATION + 10);
         clock.setMarsTime(startTime);
@@ -89,7 +55,7 @@ public class MissionBoardVehicleStepTest extends MarsSimUnitTest {
     void testEjectWorker() {
         var s = buildSettlement("Test 1");
 
-        var r = buildRoster(getContext(), s, 2, EXPLORER_ROVER, false);
+        var r = MissionTestHelper.buildRoster(getContext(), s, 2, EXPLORER_ROVER, false);
 
 		var project = new MissionVehicleProject(null, MissionType.VISIT_LANDMARK, 10, r);
         var st = new MissionBoardVehicleStep(project);
@@ -125,7 +91,7 @@ public class MissionBoardVehicleStepTest extends MarsSimUnitTest {
     void testAllBoard() {
         var s = buildSettlement("Test 1");
 
-        var r = buildRoster(getContext(), s, 2, EXPLORER_ROVER, false);
+        var r = MissionTestHelper.buildRoster(getContext(), s, 2, EXPLORER_ROVER, false);
 
 		var project = new MissionVehicleProject(null, MissionType.VISIT_LANDMARK, 10, r);
         var st = new MissionBoardVehicleStep(project);
